Phase 1 Step 8: split options_analysis into OptionsScreeningService + thin adapter
Analytics (dataclasses, fetch_security, score, build_put/call_trade, guardrails, greedy_fill, load_watchlist, is_us_listed, analyze_watchlist/symbol) move to quantcore/services/options_screening.py. options_analysis.py keeps the FastMCP server (5 thin tools), print_* presentation, and CLI main(); persistence/news route through the registry. Removes the dead store.db_path prints (Postgres OptionsStore has no db_path). 15 offline service tests; 80 green vs test DB;  47 MCP tools boot; Flask /api/health 200.
